apps/backend/requirements.txt
fastapi
uvicorn
pydantic
pyttsx3
pynput
pyautogui
pillow
pytesseract
requests
sounddevice
scipy
numpy
psutil
pywin32
wmi
pyperclip